requirements.txt
streamlit
pillow
huggingface-hub
requests